refactor: refactor to more extendable structure
