+++ node_wrappers/realtimenodes/__init__.py
@@ -1,0 +1,1 @@
+# This file intentionally left blank
