+++ node_wrappers/realtimenodes/controls/mask_controls.py
@@ -1,31 +1,8 @@
 import torch
 import numpy as np
 import cv2
-from ..base.control_base import ControlNodeBase
-
-class MaskControlMixin:
-    """Mixin providing common mask functionality"""
-    def create_circle_mask(self, height, width, center_y, center_x, size):
-        """Create a circular mask with anti-aliased edges"""
-        y, x = np.ogrid[:height, :width]
-        radius = size * min(height, width) / 2
-        dist = np.sqrt((x - center_x)**2 + (y - center_y)**2)
-        mask = np.clip(radius + 1 - dist, 0, 1).astype(np.float32)
-        return mask
-
-    def get_initial_state(self, x_pos, y_pos, size, min_size=None, max_size=None):
-        """Get initial state dictionary"""
-        state = {
-            "x": x_pos,
-            "y": y_pos,
-            "size": size,
-            "initialized": True
-        }
-        if min_size is not None:
-            state["min_size"] = min_size
-        if max_size is not None:
-            state["max_size"] = max_size
-        return state
+from ....src.realtimenodes.control_base import ControlNodeBase
+from ....src.realtimenodes.mask_controls import MaskControlMixin
 
 class RepulsiveMaskNode(ControlNodeBase, MaskControlMixin):
     """Node that maintains a mask that repulses or attracts to input masks"""
