+++ src/utils/similar_image_filter.py
